shipment_api/requirements.txt
fastapi
uvicorn
SQLAlchemy
mysql-connector-python
requests
python-dotenv
pydantic